Install postcss-cssnext without "--dev"
I previously installed the postcss-cssnext JS package with the "--dev"
flag, because that's how someone said to do it in a Github post. However
I now think it's responsible for the failures on staging and was able to
reproduce it locally. I uninstalled the --dev version and just installed
it normally and now it seems to work.