Set extract_css to true on all environments
This was causing at least one display error. It's a bit confusing to me
why they opted to make this false by default.

The same exact option is set to true in the production environment. I
opted not to remove this line, even though it's redundant, to keep as
close as possible to the webpacker defaults.