Set autoloader to :classic in environments/test.rb
Rails 6's new autoloader, zeitwerk, seems to cause some problems with
using Spring for tests. If you modify a file and re-run a spec, it'll
give the error "reloading is disabled because config.cache_classes is
true". It'll work again on the second re-run, but is inconvenient.

There's an issue on Github about it here:
https://github.com/rails/spring/issues/598. They seem to have concluded
it's a bug and the best workaround is to switch back to the classic
autoloader for now.