Refactor notified_user_ids
The old behavior was inconsistent and was probably causing users not to
receive desired notifications. For one thing, it only set
notified_user_ids when a need was updated; not when a need was created,
or a shift was created or updated. Second, it would exclude all future
notifications to the user once it was set, even if they would expect
them, such as needs or shifts being updated.

This commit changes that. Now, it will overwrite notified_user_ids
whenever a notification goes out related to the needs or shifts. It also
will not use notified_user_ids to exclude people from notifications.

This means that notified_user_ids is currently not being used at all
aside for as a field that is set when notifications go out. However, I
intend to use it in the future for the "users pending response" feature.

This commit also fixes some related specs which were passing before
whether or not the condition was being met, due to improper data setup
regarding age_ranges.