Move Reminders job creation to initializer
Probably the more appropriate place for it. Will not attempt to
re-create it if it is already found. Due to it being in an initializer,
it needs to wait for ActiveRecord to load to avoid getting a deprecation
warning related to zeitwerk.

Also re-add an empty line that was unnecessarily deleted in a prior commit.