Require all users be covid_19_vaccinated
Add a new column `users.covid_19_vaccinated`. Defaults to nil and when
a users signs in, they are forced to update their vaccination status
(yes/no). If yes, they can continue as usual. If no, an email is sent
to Sarah, and they are prevented from using the site or being assigned
shifts by coordinators, social works, and admins.