Add check for existing need
Fixes a bug when creating a new need with validation errors. Previously, if an error prevented the need from saving the date/time would not be editable. This fix checks to see if the need being created/edited exists and only allows modifying date/time if the need does not exist.